requirements.txt
fastapi
uvicorn[standard]
python-docx==0.8.11
requests
